Drop support for legacy contrib algorithms
