+++ docs/faq.rst
@@ -28,11 +28,3 @@ extract the public or private keys from a x509 certificate in PEM format.
     cert_obj = load_pem_x509_certificate(cert_str, default_backend())
     public_key = cert_obj.public_key()
     private_key = cert_obj.private_key()
-
-
-I'm using Google App Engine and can't install `cryptography`, what can I do?
-----------------------------------------------------------------------------
-
-Some platforms like Google App Engine don't allow you to install libraries
-that require C extensions to be built (like `cryptography`). If you're deploying
-to one of those environments, you should check out :ref:`legacy-deps`
