+++ pyproject.toml
@@ -28,4 +28,4 @@ use_parentheses=true
 combine_as_imports=true
 
 known_first_party="jwt"
-known_third_party=["Crypto", "cryptography", "ecdsa", "pytest", "requests_mock", "setuptools", "sphinx_rtd_theme"]
+known_third_party=["pytest", "requests_mock", "setuptools", "sphinx_rtd_theme"]
